Refactor to a common abstract auth filter.
